Converted page.tsx to component based
