+++ app/page.tsx
@@ -1,9 +1,11 @@
 "use client";
 import { processInputToCommand } from "@/model/CommandProcess";
-import Database from "@/model/Database";
-import { SetupCommandList } from "@/model/SetupCommandList";
+import Database from "@/model/CommandDataStructures/Database";
+import { SetupCommandList } from "@/model/CommandDataStructures/SetupCommandList";
 import { Command } from "@/model/Types/Types";
 import React, { useState, useRef, useEffect } from "react";
+import InputForm from "./InputForm/InputForm";
+import TerminalDisplay from "./TerminalDisplay/TerminalDisplay";
 
 export default function Home() {
 	// Accessibility synthesizer:
@@ -19,13 +21,10 @@ export default function Home() {
 	const [commands, setCommands] = useState<Command[]>([]);
 
 	// Component based state management
-	const [userInput, setUserInput] = useState<string>("");
 	const [blockInput, setBlockInput] = useState<boolean>(false);
 	const [textToSpeechRequired, setTextToSpeechRequired] =
 		useState<boolean>(false);
 	const messagesContainerRef = useRef<HTMLDivElement>(null);
-
-	const [mouseIn, setMouseIn] = useState<boolean>(false);
 
 	// Useeffects
 	useEffect(() => {
@@ -76,19 +75,16 @@ export default function Home() {
 				setBlockInput(false);
 				if (confirm("Do you want to exit?")) {
 					window.close();
+					return;
 				}
-				// window.location.href = "https://www.google.com";
+				setCommands((prevCommands) => [
+					...prevCommands,
+					{ type: "Computer", message: "Exit aborted." },
+				]);
 			}, compResponse.arg * 2000);
 		} else {
 			setBlockInput(false);
 		}
-	};
-
-	const handleInputSubmit = (e: React.FormEvent) => {
-		e.preventDefault();
-		setCommands([...commands, { type: "User", message: userInput }]);
-		processCommand(userInput);
-		setUserInput("");
 	};
 
 	const scrollToBottom = () => {
@@ -100,62 +96,10 @@ export default function Home() {
 
 	return (
 		<main className="flex min-h-screen flex-col items-center p-24 bg-black">
-			<div
-				ref={messagesContainerRef}
-				className="font-mono text-lg flex-grow rounded-t-xl w-full h-32 overflow-auto p-10
-      bg-[#2a2c3c] neuro-shadow"
-			>
-				{commands.map((command, index) => {
-					if (command.message === "") {
-						return;
-					}
-					return (
-						<div
-							key={index}
-							className={`w-full p-2 mb-2 rounded-xl ${
-								command.type === "User"
-									? "bg-black text-green-500"
-									: " bg-slate-500 text-black"
-							}`}
-						>
-							<p className="font-bold">
-								{command.type !== "Computer" ? "You: " : ""}
-							</p>
-							<p
-								className=""
-								dangerouslySetInnerHTML={{
-									__html: command.message,
-								}}
-							></p>
-						</div>
-					);
-				})}
-			</div>
-			<form onSubmit={handleInputSubmit} className="w-full">
-				<input
-					disabled={blockInput}
-					className=" rounded-bl-xl bg-white w-4/5 text-black border-b-2 border-green-500 p-2 disabled:bg-gray-800"
-					placeholder={
-						blockInput
-							? "input blocked"
-							: mouseIn
-							? "type here"
-							: "terminal ready to use, click here to start typing"
-					}
-					onFocus={() => setMouseIn(true)}
-					onBlur={() => setMouseIn(false)}
-					value={userInput}
-					onChange={(e) => setUserInput(e.target.value)}
-				/>
-				<button
-					disabled={blockInput}
-					className="rounded-br-xl bg-black w-1/5 text-green-500 border-b-2 border-green-500 p-2 hover:bg-gray-800 transition disabled:bg-green-500 disabled:text-black"
-					type="submit"
-					value="Submit"
-				>
-					{blockInput ? "running" : "Enter"}
-				</button>
-			</form>
+			<TerminalDisplay {...{ commands, messagesContainerRef }} />
+			<InputForm
+				{...{ commands, setCommands, processCommand, blockInput }}
+			/>
 		</main>
 	);
 }
